:recycle: refactor: Refactoring the params the confirmation email route
