--- src/application/candidate-user/dto/update-cadidate-user.dto.ts
@@ -1,6 +1,0 @@
-import { PartialType } from '@nestjs/mapped-types';
-import { CreateCandidateUserDto } from './create-cadidate-user.dto';
-
-export class UpdateCandidateUserDto extends PartialType(
-  CreateCandidateUserDto,
-) {}
