+++ src/application/candidate-user/dto/update-candidate-user.dto.ts
@@ -1,2 +1,6 @@
+import { PartialType } from '@nestjs/mapped-types';
+import { CreateCandidateUserDto } from './create-candidate-user.dto';
 
-export class UpdateCandidateUserDto { }
+export class UpdateCandidateUserDto extends PartialType(
+  CreateCandidateUserDto,
+) { }
