+++ src/application/candidate-user/repositories/candidate-user.repository.ts
@@ -1,5 +1,5 @@
-import { CreateCandidateUserDto } from '../dto/create-cadidate-user.dto';
-import { UpdateCandidateUserDto } from '../dto/update-cadidate-user.dto';
+import { CreateCandidateUserDto } from '../dto/create-candidate-user.dto';
+import { UpdateCandidateUserDto } from '../dto/update-candidate-user.dto';
 import { CandidateUser } from '../entities/candidate-user.entity';
 
 export abstract class CandidateUserRepository {
