--- src/common/repositories/email-confirmation/email-confirmation-sqlite.repository.ts
@@ -1,6 +1,0 @@
-import { Injectable } from "@nestjs/common";
-
-@Injectable()
-export class EmailConfirmationSqliteRepository {
-
-}
